[api-fuzzer] Move most of the logic into fuzzing_common
