+++ test/core/end2end/fuzzers/BUILD
@@ -37,6 +37,7 @@ grpc_cc_library(
     external_deps = ["absl/strings"],
     deps = [
         "api_fuzzer_proto",
+        "//test/core/event_engine/fuzzing_event_engine",
         "//:gpr",
         "//:grpc",
     ],
